Test `ImageCommentsMapper` in isolation
