--- EssentialFeed/EssentialFeedTests/Image Comments API/LoadCommentsFromRemoteUseCaseTests.swift
@@ -1,135 +1,0 @@
-//
-//  Copyright © 2018 Essential Developer. All rights reserved.
-//
-
-import XCTest
-import EssentialFeed
-
-class LoadCommentsFromRemoteUseCaseTests: XCTestCase {
-	
-	func test_load_deliversErrorOnNon2xxHTTPResponse() {
-		let (sut, client) = makeSUT()
-		
-		let samples = [199, 170, 300, 400, 500]
-		
-		samples.enumerated().forEach { index, code in
-			expect(sut, toCompleteWith: failure(.invalidData), when: {
-				let json = makeItemsJSON([])
-				client.complete(withStatusCode: code, data: json, at: index)
-			})
-		}
-	}
-	
-	func test_load_deliversErrorOn2xxHTTPResponseWithInvalidJSON() {
-		let (sut, client) = makeSUT()
-		
-		let samples = [200, 201, 250, 270, 299]
-		
-		samples.enumerated().forEach { index, code in
-			expect(sut, toCompleteWith: failure(.invalidData), when: {
-				let invalidJSON = Data("invalid json".utf8)
-				client.complete(withStatusCode: code, data: invalidJSON, at: index)
-			})
-		}
-	}
-	
-	func test_load_deliversNoItemsOn2xxHTTPResponseWithEmptyJSONList() {
-		let (sut, client) = makeSUT()
-		
-		let samples = [200, 201, 250, 270, 299]
-		
-		samples.enumerated().forEach { index, code in
-			expect(sut, toCompleteWith: .success([]), when: {
-				let emptyListJSON = makeItemsJSON([])
-				client.complete(withStatusCode: code, data: emptyListJSON, at: index)
-			})
-		}
-	}
-	
-	func test_load_deliversItemsOn2xxHTTPResponseWithJSONItems() {
-		let (sut, client) = makeSUT()
-		
-		let item1 = makeItem(
-			id: UUID(),
-			message: "a message",
-			createdAt: (Date(timeIntervalSince1970: 1619818183), "2021-04-30T21:29:43+00:00"),
-			username: "a username" )
-			
-		
-		let item2 = makeItem(
-			id: UUID(),
-			message: "another message",
-			createdAt: (Date(timeIntervalSince1970: 1619912749), "2021-05-01T23:45:49+00:00"),
-			username: "another username")
-		
-		let items = [item1.model, item2.model]
-		
-		
-		let samples = [200, 201, 250, 270, 299]
-		
-		samples.enumerated().forEach { index, code in
-			expect(sut, toCompleteWith: .success(items), when: {
-				let json = makeItemsJSON([item1.json, item2.json])
-				client.complete(withStatusCode: code, data: json, at: index)
-			})
-		}
-	}
-	
-	// MARK: - Helpers
-	
-	private func makeSUT(url: URL = URL(string: "https://a-url.com")!, file: StaticString = #filePath, line: UInt = #line) -> (sut: RemoteImageCommentsLoader, client: HTTPClientSpy) {
-		let client = HTTPClientSpy()
-		let sut = RemoteImageCommentsLoader(url: url, client: client)
-		trackForMemoryLeaks(sut, file: file, line: line)
-		trackForMemoryLeaks(client, file: file, line: line)
-		return (sut, client)
-	}
-	
-	private func failure(_ error: RemoteImageCommentsLoader.Error) -> RemoteImageCommentsLoader.Result {
-		return .failure(error)
-	}
-	
-	private func makeItem(id: UUID, message: String, createdAt:(date:Date, iso8601String: String) , username: String) -> (model: ImageComment, json: [String: Any]) {
-		let item = ImageComment(id: id, message: message, createdAt: createdAt.date, username: username)
-		
-		let json: [String: Any] = [
-			"id": id.uuidString,
-			"message": message,
-			"created_at": createdAt.iso8601String,
-			"author": [
-				"username": username
-			]
-		]
-		
-		return (item, json)
-	}
-	
-	private func makeItemsJSON(_ items: [[String: Any]]) -> Data {
-		let json = ["items": items]
-		return try! JSONSerialization.data(withJSONObject: json)
-	}
-	
-	private func expect(_ sut: RemoteImageCommentsLoader, toCompleteWith expectedResult: RemoteImageCommentsLoader.Result, when action: () -> Void, file: StaticString = #filePath, line: UInt = #line) {
-		let exp = expectation(description: "Wait for load completion")
-		
-		sut.load { receivedResult in
-			switch (receivedResult, expectedResult) {
-			case let (.success(receivedItems), .success(expectedItems)):
-				XCTAssertEqual(receivedItems, expectedItems, file: file, line: line)
-				
-			case let (.failure(receivedError as RemoteImageCommentsLoader.Error), .failure(expectedError as RemoteImageCommentsLoader.Error)):
-				XCTAssertEqual(receivedError, expectedError, file: file, line: line)
-				
-			default:
-				XCTFail("Expected result \(expectedResult) got \(receivedResult) instead", file: file, line: line)
-			}
-			
-			exp.fulfill()
-		}
-		
-		action()
-		
-		wait(for: [exp], timeout: 1.0)
-	}
-	
-}
